did stuff idk probably broken

+++ include/planetsurface.h
@@ -2,19 +2,20 @@
 #define __PLANETSURFACE_H
 #include <jsoncpp/json/json.h>
 #include <vector>
+#include <cstdint>
 class Planet;
 
 class PlanetSurface {
 public:
-    std::vector<int> tiles;
+    std::vector<uint64_t> tiles;
     bool generated = false;
     int rad = 0;
-    
+
     PlanetSurface();
     void generate(Planet * p);
 	int getType(int r, int g, int b);
     int getInitialTileType(int x, int y, Planet * p);
     Json::Value asJson();
 };
-    
+
 #endif
